Driver: Display: Nokia 5110: Use named struct instead of anonymous parameter list for template class.
Until now, not-so-long and long anonymous parameter lists were used to connect hardware drivers with the real hardware.

This is C-style as the order counts. If you swap Ce and Dc in the example above the code will compile and fail and it is hard to spot the error.

With a named struct as a single parameter each parameter has a distinguished name and is less error-prone. And the code is the documentation. Adding another parameter is simple, too.

The only downside is that default arguments cannot be defined.