[platform] Replace MAIN_FUNCTION with xpcc_main().
xpcc_main() is a much lighter wrapper around main(), and allows the
declaration of function arguments, which are passed on to main() if
the application is compiled for hosted, otherwise ignored.

This change allows renaming of function arguments and therefore the
silencing of the "unused arguments" warnings without resorting to
`MAIN_FUNCTION_NAKED`.

This change is backwards compatible, unfortunately without deprecated
warning though.